backend/requirement.txt
langgraph
langgraph.checkpoint.postgres
python-dotenv
langchain_groq
pydantic
fastapi
psycopg2-binary

